Move some stuff from InventoryRecord to BaseRecord
